Move actions to right directory and use wrangler clients where possible
